feat: add satteri support
