+++ .changeset/proud-clowns-add.md
@@ -1,0 +1,5 @@
+---
+'astro-d2': minor
+---
+
+Adds support for the Sätteri Markdown processor.
