+++ eslint.config.mjs
@@ -1,7 +1,7 @@
 import hideoo from '@hideoo/eslint-config'
 
 export default hideoo({
-  files: ['packages/astro-d2/tests/remark-d2js.test.ts'],
+  files: ['packages/astro-d2/tests/markdown-d2js.test.ts'],
   rules: {
     '@typescript-eslint/unbound-method': 'off',
   },
