+++ packages/astro-d2/package.json
@@ -17,6 +17,7 @@
     "@terrastruct/d2": "^0.1.33",
     "hast-util-from-html": "^2.0.3",
     "hast-util-to-html": "^9.0.5",
+    "satteri": "^0.9.4",
     "unist-util-visit": "^5.1.0"
   },
   "devDependencies": {
